# alumni-webapp/api-docs/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:21001","description":"Generated server url"}],"tags":[{"name":"Employee"},{"name":"TokenProvider"},{"name":"Payroll"}],"paths":{"/api/tokens":{"post":{"tags":["TokenProvider"],"operationId":"token","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK"}}}},"/api/employees/{zalarisId}/personal-data":{"get":{"tags":["Employee"],"operationId":"getPersonalData","parameters":[{"name":"zalarisId","in":"path","required":true,"schema":{"pattern":"\\d{3}-\\d{8}","type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PersonalDataModel"}}}}}}},"/api/employees/{zalarisId}/payroll-results":{"get":{"tags":["Payroll"],"operationId":"getPayrollResults","parameters":[{"name":"zalarisId","in":"path","required":true,"schema":{"pattern":"\\d{3}-\\d{8}","type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PayrollResultsModel"}}}}}}},"/api/employees/{zalarisId}/payroll-results/{sequenceId}/payslip":{"get":{"tags":["Payroll"],"operationId":"getPayslip","parameters":[{"name":"zalarisId","in":"path","required":true,"schema":{"pattern":"\\d{3}-\\d{8}","type":"string"}},{"name":"sequenceId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"application/pdf":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/employees/employment/history":{"get":{"tags":["Employee"],"operationId":"getEmploymentHistory","parameters":[{"name":"ssid","in":"query","required":false,"schema":{"type":"string"}},{"name":"zalarisId","in":"query","required":false,"schema":{"pattern":"\\d{3}-\\d{8}","type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmploymentHistoryModel"}}}}}}}},"components":{"schemas":{"Error":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"}}},"ErrorResponse":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"timestamp":{"type":"string","format":"date-time"},"path":{"type":"string"},"errors":{"type":"array","items":{"$ref":"#/components/schemas/Error"}}}},"AdditionalData":{"type":"object","properties":{"personalIdentifier":{"type":"string"},"birthDate":{"type":"string","format":"date"},"language":{"type":"string"},"nationality":{"type":"string"},"maritalStatusKey":{"type":"string"},"nrOfChildren":{"type":"string"}}},"AdditionalPaymentDeduction":{"type":"object","properties":{"amount":{"type":"string"},"type":{"type":"string"},"currency":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"}}},"Address":{"type":"object","properties":{"addressType":{"type":"string"},"street":{"type":"string"},"street2ndLine":{"type":"string"},"houseNumber":{"type":"string"},"apartment":{"type":"string"},"postalCode":{"type":"string"},"city":{"type":"string"},"district":{"type":"string"},"state":{"type":"string"},"country":{"type":"string"},"phoneNumber":{"type":"string"}}},"Administrator":{"type":"object","properties":{"group":{"type":"string"},"timeRecording":{"type":"string"},"hr":{"type":"string"},"payroll":{"type":"string"},"supervisor":{"type":"string"}}},"BankDetails":{"type":"object","properties":{"bankDetailsType":{"type":"string"},"payee":{"type":"string"},"postCode":{"type":"string"},"city":{"type":"string"},"bankCountryKey":{"type":"string"},"bankKey":{"type":"string"},"bankNumber":{"type":"string"},"bankControlKey":{"type":"string"},"swift":{"type":"string"},"paymentMethod":{"type":"string"},"currency":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"}}},"BasicPayGroups":{"type":"object","properties":{"personnel":{"$ref":"#/components/schemas/Personnel"},"payscale":{"$ref":"#/components/schemas/Payscale"},"payWages":{"type":"array","items":{"$ref":"#/components/schemas/Paywage"}}}},"Communication":{"type":"object","properties":{"type":{"type":"string"},"communicationId":{"type":"string"}}},"EnterpriseStructure":{"type":"object","properties":{"companyCode":{"type":"string"},"companyName":{"type":"string"},"personnelArea":{"type":"string"},"personnelSubarea":{"type":"string"},"costCenter":{"type":"string"},"costCenterName":{"type":"string"}}},"Name":{"type":"object","properties":{"formOfAddressKey":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"nickName":{"type":"string"},"genderKey":{"type":"string"},"initials":{"type":"string"}}},"OrganizationalAssignmentGroups":{"type":"object","properties":{"personnel":{"$ref":"#/components/schemas/Personnel"},"enterpriseStructure":{"$ref":"#/components/schemas/EnterpriseStructure"},"personnelStructure":{"$ref":"#/components/schemas/PersonnelStructure"},"organizationalPlan":{"$ref":"#/components/schemas/OrganizationalPlan"},"administrator":{"$ref":"#/components/schemas/Administrator"}}},"OrganizationalPlan":{"type":"object","properties":{"position":{"type":"string"},"job":{"type":"string"},"organizationUnit":{"type":"string"},"organizationKey":{"type":"string"}}},"Payscale":{"type":"object","properties":{"subtype":{"type":"string"},"reason":{"type":"string"},"type":{"type":"string"},"area":{"type":"string"},"group":{"type":"string"},"capacityUtilizationLevel":{"type":"string"},"workingHoursPerPayrollPeriod":{"type":"string"},"annualSalary":{"type":"string"},"annualSalaryCurrencyKey":{"type":"string"}}},"Paywage":{"type":"object","properties":{"type":{"type":"string"},"amount":{"type":"string"},"unit":{"type":"string"}}},"PersonalDataModel":{"type":"object","properties":{"ssid":{"type":"string"},"zalarisId":{"type":"string"},"orgAssignment":{"$ref":"#/components/schemas/OrganizationalAssignmentGroups"},"personalInformation":{"$ref":"#/components/schemas/PersonalInformationGroups"},"addresses":{"type":"array","items":{"$ref":"#/components/schemas/Address"}},"communications":{"type":"array","items":{"$ref":"#/components/schemas/Communication"}},"basicPay":{"$ref":"#/components/schemas/BasicPayGroups"},"recurringPaymentDeductions":{"type":"array","items":{"$ref":"#/components/schemas/RecurringPaymentDeduction"}},"additionalPaymentDeductions":{"type":"array","items":{"$ref":"#/components/schemas/AdditionalPaymentDeduction"}},"bankDetails":{"type":"array","items":{"$ref":"#/components/schemas/BankDetails"}}}},"PersonalInformationGroups":{"type":"object","properties":{"personnel":{"$ref":"#/components/schemas/Personnel"},"name":{"$ref":"#/components/schemas/Name"},"additionalData":{"$ref":"#/components/schemas/AdditionalData"}}},"Personnel":{"type":"object","properties":{"personnelNo":{"type":"string"},"changedAt":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"}}},"PersonnelStructure":{"type":"object","properties":{"payrollArea":{"type":"string"},"contract":{"type":"string"},"employeeGroup":{"type":"string"},"employeeSubgroup":{"type":"string"}}},"RecurringPaymentDeduction":{"type":"object","properties":{"amount":{"type":"string"},"type":{"type":"string"},"currency":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"}}},"BankTransfer":{"type":"object","properties":{"wageType":{"type":"string"},"hrPayrollAmount":{"type":"string"},"bankCountryKey":{"type":"string"},"bankNumber":{"type":"string"},"bankAccountNumber":{"type":"string"},"transferDate":{"type":"string"},"currency":{"type":"string"}}},"PayrollResult":{"type":"object","properties":{"sequence":{"type":"string"},"area":{"type":"string"},"inPeriod":{"type":"string"},"forPeriod":{"type":"string"},"startPayroll":{"type":"string"},"endPayroll":{"type":"string"},"companyCode":{"type":"string"},"payDate":{"type":"string"},"amountPaid":{"type":"string"},"currency":{"type":"string"},"payrollResultsDetails":{"type":"array","items":{"$ref":"#/components/schemas/PayrollResultEntry"}},"cumulativePayrollResults":{"type":"array","items":{"$ref":"#/components/schemas/PayrollResultEntry"}},"bankTransfers":{"type":"array","items":{"$ref":"#/components/schemas/BankTransfer"}}}},"PayrollResultEntry":{"type":"object","properties":{"wageType":{"type":"string"},"hrPayrollNumber":{"type":"string"},"hrPayrollAmount":{"type":"string"}}},"PayrollResultsModel":{"type":"object","properties":{"ssid":{"type":"string"},"zalarisId":{"type":"string"},"payrollResults":{"type":"array","items":{"$ref":"#/components/schemas/PayrollResult"}}}},"Employment":{"type":"object","properties":{"zalarisId":{"type":"string"},"companyCode":{"type":"string"},"companyName":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"}}},"EmploymentHistoryModel":{"type":"object","properties":{"ssid":{"type":"string"},"employments":{"type":"array","items":{"$ref":"#/components/schemas/Employment"}}}}}}}